#12: EL-Providers shall not wrap exceptions from invoked methods.
For ELProviderExceptions, no stack traces are reported, since errors in
EL Expressions are "human". 
Unfortunately, if there is some bug in japkit (for example in
TypeExtensions), the Java EL implementation will wrap it into
ElException.
So if this is then wrapped as ELProviderException there will be no
stacktrace for the bug.
Thus, there sahll be no ELProviderException wrapping, if the ElException
has a cause that is NOT an ElException.